chore(next-config): next 15に対応
related #35
